Version 1.1: Refactor bot service
