+++ config.py
@@ -7,3 +7,4 @@ load_dotenv()
 API_ID = environ.get('API_ID')
 API_HASH = environ.get('API_HASH')
 SESSION_STRING = environ.get('SESSION_STRING')
+TOKEN = environ.get('TOKEN')
