--- r.txt
@@ -1,5 +1,0 @@
-telebot
-python-dotenv
-telethon
-openai
-aiohttp
