+++ requirements.txt
@@ -1,0 +1,8 @@
+python-dotenv
+telethon
+openai
+aiohttp
+aiogram
+asyncpg
+aiofiles
+aiohttp
